Import py33-Pmw-2.0.0rc1 as x11/py-Pmw2.
Pmw is a toolkit for building high-level compound widgets in Python
using the Tkinter module.

It consists of a set of base classes and a library of flexible and
extensible megawidgets built on this foundation. These megawidgets
include notebooks, comboboxes, selection widgets, paned widgets,
scrolled widgets and dialog windows.

This is the python-3.x version of the module.