Import py27-pbr-0.8.2 as devel/py-pbr.
PBR is a library that injects some useful and sensible default behaviors into
your setuptools run. It started off life as the chunks of code that were copied
between all of the OpenStack projects. Around the time that OpenStack hit 18
different projects each with at least 3 active branches, it seemed like a good
time to make that code into a proper reusable library.

PBR is only mildly configurable. The basic idea is that there's a decent way to
run things and if you do, you should reap the rewards, because then it's simple
and repeatable. If you want to do things differently, cool! But you've already
got the power of Python at your fingertips, so you don't really need PBR.

PBR builds on top of the work that d2to1 started to provide for declarative
configuration. d2to1 is itself an implementation of the ideas behind distutils2.
Although distutils2 is now abandoned in favor of work towards PEP 426 and
Metadata 2.0, declarative config is still a great idea and specifically
important in trying to distribute setup code as a library when that library
itself will alter how the setup is processed. As Metadata 2.0 and other modern
Python packaging PEPs come out, PBR aims to support them as quickly as possible.