Update to Poly/ML 5.5.2
Most of the changes are bug fixes.  One major fix was to the
intermediate code optimiser reported back in September.
Other changes are the addition of PrettyStringWithWidth and
PrettyLineBreak constructors to the PolyML.pretty datatype,
Thread.Thread.numPhysicalProcessors to return the number of
physical processors and printing ?.t in circumstances where t is
bound to a different type.  The debugger has been extended to
include structures and the arguments to functors.